refactor: move malware detection constants to a dedicated file for better organization
