+++ plugins/wazuh-analysis/public/groups/endpoint-security/applications.ts
@@ -5,12 +5,15 @@ import {
   AppNavLinkStatus,
   AppUpdater,
 } from '../../../../../src/core/public';
-import { MALWARE_DETECTION_ID, MALWARE_DETECTION_TITLE } from './constants';
 import {
   CONFIGURATION_ASSESSMENT_ID,
   CONFIGURATION_ASSESSMENT_TITLE,
 } from './apps/configuration-assesment/constants';
 import { FIM_ID, FIM_TITLE } from './apps/fim/constants';
+import {
+  MALWARE_DETECTION_ID,
+  MALWARE_DETECTION_TITLE,
+} from './apps/malware-detection/constants';
 
 export function getEndpointSecurityApps(updater$?: Subject<AppUpdater>): App[] {
   return [
