+++ plugins/wazuh-analysis/public/groups/endpoint-security/apps/malware-detection/constants.ts
@@ -1,0 +1,15 @@
+import { i18n } from '@osd/i18n';
+import { PLUGIN_ID } from '../../../../../common/constants';
+import { buildSubAppId } from '../../../../utils';
+import { ENDPOINT_SECURITY_ID } from '../../constants';
+
+export const MALWARE_DETECTION_ID = buildSubAppId(
+  ENDPOINT_SECURITY_ID,
+  'malware_detection',
+);
+export const MALWARE_DETECTION_TITLE = i18n.translate(
+  `${PLUGIN_ID}.category.${MALWARE_DETECTION_ID}`,
+  {
+    defaultMessage: 'Malware Detection',
+  },
+);
