+++ plugins/wazuh-analysis/public/groups/endpoint-security/apps/malware-detection/malware-detection-app.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { AppMountParameters } from 'opensearch-dashboards/public';
-import { MALWARE_DETECTION_TITLE } from '../../constants';
+import { MALWARE_DETECTION_TITLE } from './constants';
 
 interface MalwareDetectionAppProps {
   params: AppMountParameters;
