+++ plugins/wazuh-analysis/public/groups/endpoint-security/constants.ts
@@ -1,6 +1,5 @@
 import { i18n } from '@osd/i18n';
 import { PLUGIN_ID } from '../../../common/constants';
-import { buildSubAppId } from '../../utils';
 
 export const ENDPOINT_SECURITY_ID = 'endpoint_security';
 export const ENDPOINT_SECURITY_TITLE = i18n.translate(
@@ -14,15 +13,5 @@ export const ENDPOINT_SECURITY_DESCRIPTION = i18n.translate(
   {
     defaultMessage:
       'Advanced monitoring and protection for devices against security threats.',
-  },
-);
-export const MALWARE_DETECTION_ID = buildSubAppId(
-  ENDPOINT_SECURITY_ID,
-  'malware_detection',
-);
-export const MALWARE_DETECTION_TITLE = i18n.translate(
-  `${PLUGIN_ID}.category.${MALWARE_DETECTION_ID}`,
-  {
-    defaultMessage: 'Malware Detection',
   },
 );
